♻️ refactor: Extract `TimerState` to a dedicated type and enhance type safety in timer logic
