+++ app/const/clientShell.ts
@@ -17,6 +17,5 @@ const stepKindToSection = (kind: string): string | null => {
 
 export {
     QUESTION_KINDS,
-    SECTIONS,
-    stepKindToSection,
-}
+    SECTIONS, stepKindToSection
+};
