+++ app/types/timerStore.ts
@@ -1,0 +1,18 @@
+import {Section} from "@/types/clientShell";
+
+interface TimerState {
+    globalTimeLeft: number;
+    sectionTimeLeft: number;
+    currentSection: Section | null;
+    isRunning: boolean;
+
+
+    sectionElapsed: Record<Section, number>;
+
+    tick: () => void;
+    startSection: (section: Section) => void;
+    start: () => void;
+    pause: () => void;
+}
+
+export type {TimerState};
